Add an Uncached Asset controller for admin
