+++ admin/app/controllers/UncachedAssets.scala
@@ -1,0 +1,4 @@
+package controllers.admin
+
+object UncachedAssets extends controllers.PublicAssets
+object UncachedWebAssets extends controllers.PublicAssets
