+++ common/app/controllers/PublicAssets.scala
@@ -1,0 +1,7 @@
+package controllers
+
+trait PublicAssets extends play.api.mvc.Controller {
+  def at(file: String, relativePath: String = "") = model.NoCache {
+    controllers.Assets.at("/public", relativePath + file)
+  }
+}
