+++ dev-build/conf/routes
@@ -6,6 +6,8 @@ GET         /surveys/*file                                                      
 GET         /commercial/test-page                                     												 controllers.commercial.CreativeTestPage.allComponents(k: List[String])
 
 # For dev machines
+GET         /assets/admin/lib/*file                                                                                  controllers.admin.UncachedWebAssets.at(file)
+GET         /assets/admin/*file                                                                                      controllers.admin.UncachedAssets.at(file, relativePath="lib/admin/")
 GET         /assets/internal/*file                                                                                   controllers.Assets.at(path="/public", file)
 GET         /assets/*path                                                                                            dev.DevAssetsController.at(path)
