+++ facia-tool/app/controllers/UncachedAssets.scala
@@ -1,10 +1,3 @@
 package controllers.faciatool
 
-import actions.NoCache
-import play.api.mvc.Controller
-
-object UncachedAssets extends Controller {
-  def at(file: String, relativePath: String = "") = NoCache {
-    controllers.Assets.at("/public", relativePath + file)
-  }
-}
+object UncachedAssets extends controllers.PublicAssets
